use typescript for requestsReducer
